Add files via upload
A Different approach to loading data from the CSV File, now employing the use of a dictionary as appose to a list in the inital loading of the customer, then allowing for the removal of the need for a purchase list and instead loading the object and appending to the relevant customer utilising the customer ID function as a key for the dict. Eliminates the need for certain processing and creating on objects in the code albeit not by much.